Re-org folder structures (also split RAM and Tag2Text model files), getting ready for packaging.
