--- configs/q2l_config.json
@@ -1,22 +1,0 @@
-{
-    "architectures": [
-      "BertModel"
-    ],
-    "attention_probs_dropout_prob": 0.1,
-    "hidden_act": "gelu",
-    "hidden_dropout_prob": 0.1,
-    "hidden_size": 768,
-    "initializer_range": 0.02,
-    "intermediate_size": 3072,
-    "layer_norm_eps": 1e-12,
-    "max_position_embeddings": 512,
-    "model_type": "bert",
-    "num_attention_heads": 4,
-    "num_hidden_layers": 2,
-    "pad_token_id": 0,
-    "type_vocab_size": 2,
-    "vocab_size": 30522,
-    "encoder_width": 768,
-    "add_cross_attention": true,
-    "add_tag_cross_attention": false
-  }
