+++ inference_ram_openset.py
@@ -10,9 +10,9 @@ import torch
 import torchvision.transforms as transforms
 
 from PIL import Image
-from models.tag2text import ram
+from ram.models import ram
 
-from models.openset_utils import build_openset_label_embedding
+from ram.utils import build_openset_label_embedding
 from torch import nn
 
 parser = argparse.ArgumentParser(
