+++ inference_tag2text.py
@@ -10,7 +10,7 @@ import torch
 import torchvision.transforms as transforms
 
 from PIL import Image
-from models.tag2text import tag2text_caption
+from ram.models import tag2text_caption
 
 parser = argparse.ArgumentParser(
     description='Tag2Text inferece for tagging and captioning')
