+++ ram/__init__.py
@@ -1,0 +1,1 @@
+from .inference import inference_tag2text, inference_ram, inference_ram_openset
