+++ ram/configs/swin/config_swinL_384.json
@@ -1,0 +1,9 @@
+{
+    "ckpt": "pretrain_model/swin_large_patch4_window12_384_22k.pth",
+    "vision_width": 1536,
+    "image_res": 384,
+    "window_size": 12,
+    "embed_dim": 192,
+    "depths": [ 2, 2, 18, 2 ],
+    "num_heads": [ 6, 12, 24, 48 ]
+  }
