+++ ram/models/__init__.py
@@ -1,0 +1,2 @@
+from .ram import ram
+from .tag2text import tag2text_caption
