+++ ram/utils/__init__.py
@@ -1,0 +1,1 @@
+from .openset_utils import build_openset_label_embedding
